工具类更新

+++ Swift/DCUtils/DCUtils.swift
@@ -8,21 +8,16 @@
 
 import UIKit
 class DCUtils: NSObject {
-    
     //判断版本号
-    class  func isNewVersion() -> Bool {
-        //  var flag = false
-        //已经存储的
-        let currentVersion = dc_bundleVersion()
-        let version = Double(currentVersion)!
+    class  func dc_isNeedUpdate(ver:String) -> Bool {
+        let preVer = dc_bundleVersion()
+        let curVer = ver
         //比较两个版本
-        
-        let sandboxVersionKey = "sandboxVersionKey"
-        let sandboxVersion = UserDefaults.standard.double(forKey: sandboxVersionKey)
-        
-        UserDefaults.standard.set(version, forKey: sandboxVersionKey)
-        
-        return version > sandboxVersion
+        let preArr = preVer.split(separator: ".")
+        let curArr = curVer.split(separator: ".")
+        let pV:Int = Int(preArr.joined(separator: "")) ?? 0
+        let cV:Int = Int(curArr.joined(separator: "")) ?? 0
+        return cV > pV
     }
     //获取版本号
     class  func dc_bundleVersion() -> String {
@@ -33,20 +28,82 @@ class DCUtils: NSObject {
     class func dc_jumpToAppleStore() {
         let string = "itms-apps://itunes.apple.com/cn/app/id1043010693?mt=8"
         if #available(iOS 10.0, *) {
-        UIApplication.shared.open(URL.init(string: string)!, options: [UIApplicationOpenURLOptionUniversalLinksOnly : true], completionHandler: nil)
+            UIApplication.shared.open(URL.init(string: string)!, options: [UIApplication.OpenExternalURLOptionsKey.universalLinksOnly : true], completionHandler: nil)
          }else {
             UIApplication.shared.openURL(URL.init(string: string)!)
         }
     }
-    //callPhone
+    ///打电话
     class func dc_callNumber(phone:String) {
         let telUrl = "telprompt:\(phone)"
         if #available(iOS 10.0, *) {
-            UIApplication.shared.open(URL.init(string: telUrl)!, options: [UIApplicationOpenURLOptionUniversalLinksOnly : true], completionHandler: nil)
+            UIApplication.shared.open(URL.init(string: telUrl)!, options: [UIApplication.OpenExternalURLOptionsKey.universalLinksOnly : true], completionHandler: nil)
         }else {
             UIApplication.shared.openURL(URL.init(string: telUrl)!)
         }
         
+    }
+    /// 验证手机号
+    class func dc_verifyPhoneNumber(phoneNumber:String) -> Bool {
+        let  MOBILE = "^1[0-9]{10}$";
+        let regextestmobile = NSPredicate(format: "SELF MATCHES %@",MOBILE)
+        return  regextestmobile.evaluate(with: phoneNumber)
+    }
+    /// 验证密码
+    class func dc_verifyPassword(password:String) -> Bool {
+        let  pwd = "^(?![0-9]+$)(?![a-zA-Z]+$)[a-zA-Z0-9]{8,16}"
+        let regextestmobile = NSPredicate(format: "SELF MATCHES %@",pwd)
+        return  regextestmobile.evaluate(with: password)
+    }
+    /// 验证验证码
+    class func dc_verifyVerCode(code:String) -> Bool {
+        let pwd  = "^[0-9]{6}$"
+        let regextestmobile = NSPredicate(format: "SELF MATCHES %@",pwd)
+        return  regextestmobile.evaluate(with: code)
+    }
+    /// 验证身份证号
+    class func dc_verifyID(ID:String) -> Bool {
+        let pwd = "^(\\d{14}|\\d{17})(\\d|[xX])$"
+        let regextestmobile = NSPredicate(format: "SELF MATCHES %@",pwd)
+        return  regextestmobile.evaluate(with: ID)
+    }
+    ///返回密文手机号
+    class func dc_hidePhone(phoneNumber:String) -> String {
+        if self.dc_verifyPhoneNumber(phoneNumber: phoneNumber) {
+            let phone = NSString.init(string: phoneNumber)
+            return phone.replacingCharacters(in: NSMakeRange(3, 5), with:"*****")
+        }
+        return "*****";
+    }
+    ///返回对应周几
+    class func dc_weekDay(index:Int) -> String {
+        var week = "未知";
+        switch (index) {
+        case 0:
+            week = "周日";
+            break;
+        case 1:
+            week =  "周一";
+            break;
+        case 2:
+            week =  "周二";
+            break;
+        case 3:
+            week =  "周三";
+            break;
+        case 4:
+            week =  "周四";
+            break;
+        case 5:
+            week =  "周五";
+            break;
+        case 6:
+            week =  "周六";
+            break;
+        default:
+            break;
+        }
+        return week
     }
     //MARK:- UserDefault
     //保存 字典
@@ -56,7 +113,6 @@ class DCUtils: NSObject {
         }
     }
     class  func dc_setObject(data:Any,key:String) {
-       
        
         UserDefaults.standard.set(data, forKey: key)
         
@@ -70,10 +126,11 @@ class DCUtils: NSObject {
     }
     // MARK: - cookies
     class func dc_saveCookies() {
-        
-        let cookiesData = NSKeyedArchiver.archivedData(withRootObject: HTTPCookieStorage.shared.cookies as Any)
-        for cookie in HTTPCookieStorage.shared.cookies! {
-            self.dc_setObject(data: cookie.value, key: cookie.name)
+        var cookiesData:Data;
+        if #available(iOS 12.0, *) {
+            cookiesData = try! NSKeyedArchiver.archivedData(withRootObject: HTTPCookieStorage.shared.cookies as Any, requiringSecureCoding: false)
+        }else {
+            cookiesData = NSKeyedArchiver.archivedData(withRootObject: HTTPCookieStorage.shared.cookies as Any)
             
         }
         self.dc_setObject(data: cookiesData, key: "DCCookies")
@@ -82,21 +139,21 @@ class DCUtils: NSObject {
         if (dc_getObjectForKey(key: "DCCookies") == nil) {
             return
         }
-        
-        let cookies = NSKeyedUnarchiver.unarchiveObject(with: self.dc_getObjectForKey(key: "DCCookies") as! Data) as! [HTTPCookie]
+        let  cookies = NSKeyedUnarchiver.unarchiveObject(with: self.dc_getObjectForKey(key: "DCCookies") as! Data) as! [HTTPCookie]
         let cookieStorage = HTTPCookieStorage.shared
         for cookie in cookies {
-            
             cookieStorage.setCookie(cookie)
         }
         
     }
     class func dc_deleteCookies() {
+        if (dc_getObjectForKey(key: "DCCookies") == nil) {
+            return
+        }
         let cookieStorage = HTTPCookieStorage.shared
         for cookie in  HTTPCookieStorage.shared.cookies! {
             cookieStorage.deleteCookie(cookie)
         }
-        
+        self.dc_deleteObject(key: "DCCookies")
     }
-
 }
